Move loadThoughts to action-creator.
